requirements.txt
pandas>=1.3.0 
fastapi
uvicorn[standard]
pydantic-ai
openai
streamlit
requests
psycopg2-binary
python-dotenv
asyncpg
sqlalchemy
